+++ modules/controllers/content-controller.ts
@@ -1,3 +1,4 @@
+import { syncBlockedBuffers } from '../actions/content/sync-erc4626-blocked-buffers';
 import { syncErc4626Reviews } from '../actions/content/sync-erc4626-reviews';
 import { syncHookReviews } from '../actions/content/sync-hook-reviews';
 import { syncRateProviderReviews } from '../actions/content/sync-rate-provider-reviews';
@@ -13,7 +14,8 @@ export function ContentController(tracer?: any) {
         async syncHookReviews() {
             return await syncHookReviews();
         },
-        async syncErc4626Reviews() {
+        async syncErc4626Data() {
+            await syncBlockedBuffers();
             return await syncErc4626Reviews();
         },
         async syncCategories() {